サマリー画面のレイアウトを修正
